+++ eslint.config.mjs
@@ -162,6 +162,12 @@ export default tseslint.config(
           ignorePrimitives: true,
         },
       ],
+      '@typescript-eslint/no-require-imports': [
+        'error',
+        {
+          allow: ['/package\\.json$'],
+        },
+      ],
 
       //
       // Internal repo rules
@@ -195,6 +201,7 @@ export default tseslint.config(
         { commentPattern: '.*intentional fallthrough.*' },
       ],
       'one-var': ['error', 'never'],
+      'prefer-object-has-own': 'error',
 
       //
       // eslint-plugin-eslint-comment