+++ packages/rule-tester/src/RuleTester.ts
@@ -20,6 +20,7 @@ import { Linter } from '@typescript-eslint/utils/ts-eslint';
 // we intentionally import from eslint here because we need to use the same class
 // that ESLint uses, not our custom override typed version
 import { SourceCode } from 'eslint';
+import stringify from 'json-stable-stringify-without-jsonify';
 import merge from 'lodash.merge';
 
 import { TestFramework } from './TestFramework';
@@ -40,6 +41,7 @@ import { getRuleOptionsSchema } from './utils/getRuleOptionsSchema';
 import { hasOwnProperty } from './utils/hasOwnProperty';
 import { getPlaceholderMatcher, interpolate } from './utils/interpolate';
 import { isReadonlyArray } from './utils/isReadonlyArray';
+import { isSerializable } from './utils/serialization';
 import * as SourceCodeFixer from './utils/SourceCodeFixer';
 import {
   emitLegacyRuleAPIWarning,
@@ -115,6 +117,8 @@ function getUnsubstitutedMessagePlaceholders(
 export class RuleTester extends TestFramework {
   readonly #testerConfig: TesterConfigWithDefaults;
   readonly #rules: Record<string, AnyRuleCreateFunction | AnyRuleModule> = {};
+  // `configType` is hardcoded to eslintrc until switch to flat config is complete
+  // We should look to support both eslintrc and flat config for consumers of RuleTester.
   readonly #linter: Linter = new Linter({ configType: 'eslintrc' });
 
   /**
@@ -142,6 +146,7 @@ export class RuleTester extends TestFramework {
       try {
         // instead of creating a hard dependency, just use a soft require
         // a bit weird, but if they're using this tooling, it'll be installed
+        // eslint-disable-next-line @typescript-eslint/no-require-imports
         const parser = require(TYPESCRIPT_ESLINT_PARSER) as typeof ParserType;
         parser.clearCaches();
       } catch {
@@ -253,16 +258,19 @@ export class RuleTester extends TestFramework {
       if (test.parser === TYPESCRIPT_ESLINT_PARSER) {
         throw new Error(DUPLICATE_PARSER_ERROR_MESSAGE);
       }
-      return {
-        ...test,
-        filename: test.filename || getFilename(test.parserOptions),
-        parserOptions: {
-          // Re-running simulates --fix mode, which implies an isolated program
-          // (i.e. parseAndGenerateServicesCalls[test.filename] > 1).
-          disallowAutomaticSingleRunInference: true,
-          ...test.parserOptions,
-        },
-      };
+      if (!test.filename) {
+        return {
+          ...test,
+          filename: getFilename(test.parserOptions),
+          parserOptions: {
+            // Re-running simulates --fix mode, which implies an isolated program
+            // (i.e. parseAndGenerateServicesCalls[test.filename] > 1).
+            disallowAutomaticSingleRunInference: true,
+            ...test.parserOptions,
+          },
+        };
+      }
+      return test;
     };
 
     const normalizedTests = {
@@ -396,6 +404,9 @@ export class RuleTester extends TestFramework {
       );
     }
 
+    const seenValidTestCases = new Set<string>();
+    const seenInvalidTestCases = new Set<string>();
+
     if (typeof rule === 'function') {
       emitLegacyRuleAPIWarning(ruleName);
     }
@@ -445,7 +456,12 @@ export class RuleTester extends TestFramework {
               return valid.name;
             })();
             constructor[getTestMethod(valid)](sanitize(testName), () => {
-              this.#testValidTemplate(ruleName, rule, valid);
+              this.#testValidTemplate(
+                ruleName,
+                rule,
+                valid,
+                seenValidTestCases,
+              );
             });
           });
         });
@@ -461,7 +477,12 @@ export class RuleTester extends TestFramework {
               return invalid.name;
             })();
             constructor[getTestMethod(invalid)](sanitize(name), () => {
-              this.#testInvalidTemplate(ruleName, rule, invalid);
+              this.#testInvalidTemplate(
+                ruleName,
+                rule,
+                invalid,
+                seenInvalidTestCases,
+              );
             });
           });
         });
@@ -486,6 +507,8 @@ export class RuleTester extends TestFramework {
     outputs: string[];
     beforeAST: TSESTree.Program;
     afterAST: TSESTree.Program;
+    config: RuleTesterConfig;
+    filename?: string;
   } {
     let config: TesterConfigWithDefaults = merge({}, this.#testerConfig);
     let code;
@@ -565,6 +588,7 @@ export class RuleTester extends TestFramework {
 
     this.#linter.defineParser(
       config.parser,
+      // eslint-disable-next-line @typescript-eslint/no-require-imports
       wrapParser(require(config.parser) as Parser.ParserModule),
     );
 
@@ -666,6 +690,8 @@ export class RuleTester extends TestFramework {
     } while (fixedResult.fixed && passNumber < 10);
 
     return {
+      config,
+      filename,
       messages: initialMessages,
       outputs,
       // is definitely assigned within the `rule-tester/validate-ast` rule
@@ -688,6 +714,7 @@ export class RuleTester extends TestFramework {
     ruleName: string,
     rule: RuleModule<MessageIds, Options>,
     itemIn: ValidTestCase<Options> | string,
+    seenValidTestCases: Set<string>,
   ): void {
     const item: ValidTestCase<Options> =
       typeof itemIn === 'object' ? itemIn : { code: itemIn };
@@ -702,6 +729,8 @@ export class RuleTester extends TestFramework {
         "Optional test case property 'name' must be a string",
       );
     }
+
+    checkDuplicateTestCase(item, seenValidTestCases);
 
     const result = this.runRuleForItem(ruleName, rule, item);
     const messages = result.messages;
@@ -730,6 +759,7 @@ export class RuleTester extends TestFramework {
     ruleName: string,
     rule: RuleModule<MessageIds, Options>,
     item: InvalidTestCase<MessageIds, Options>,
+    seenInvalidTestCases: Set<string>,
   ): void {
     assert.ok(
       typeof item.code === 'string',
@@ -749,6 +779,8 @@ export class RuleTester extends TestFramework {
     if (Array.isArray(item.errors) && item.errors.length === 0) {
       assert.fail('Invalid cases must have at least one error');
     }
+
+    checkDuplicateTestCase(item, seenInvalidTestCases);
 
     const ruleHasMetaMessages =
       hasOwnProperty(rule, 'meta') && hasOwnProperty(rule.meta, 'messages');
@@ -1064,6 +1096,25 @@ export class RuleTester extends TestFramework {
                     [actualSuggestion],
                   ).output;
 
+                  // Verify if suggestion fix makes a syntax error or not.
+                  const errorMessageInSuggestion = this.#linter
+                    .verify(
+                      codeWithAppliedSuggestion,
+                      result.config,
+                      result.filename,
+                    )
+                    .find(m => m.fatal);
+
+                  assert(
+                    !errorMessageInSuggestion,
+                    [
+                      'A fatal parsing error occurred in suggestion fix.',
+                      `Error: ${errorMessageInSuggestion?.message}`,
+                      'Suggestion output:',
+                      codeWithAppliedSuggestion,
+                    ].join('\n'),
+                  );
+
                   assert.strictEqual(
                     codeWithAppliedSuggestion,
                     expectedSuggestion.output,
@@ -1134,6 +1185,30 @@ export class RuleTester extends TestFramework {
  */
 function assertASTDidntChange(beforeAST: unknown, afterAST: unknown): void {
   assert.deepStrictEqual(beforeAST, afterAST, 'Rule should not modify AST.');
+}
+
+/**
+ * Check if this test case is a duplicate of one we have seen before.
+ */
+function checkDuplicateTestCase(
+  item: unknown,
+  seenTestCases: Set<unknown>,
+): void {
+  if (!isSerializable(item)) {
+    /*
+     * If we can't serialize a test case (because it contains a function, RegExp, etc), skip the check.
+     * This might happen with properties like: options, plugins, settings, languageOptions.parser, languageOptions.parserOptions.
+     */
+    return;
+  }
+
+  const serializedTestCase = stringify(item);
+
+  assert(
+    !seenTestCases.has(serializedTestCase),
+    'detected duplicate test case',
+  );
+  seenTestCases.add(serializedTestCase);
 }
 
 /**