+++ packages/typescript-eslint/src/configs/recommended-type-checked-only.ts
@@ -37,7 +37,6 @@ export default (
       '@typescript-eslint/no-unsafe-member-access': 'error',
       '@typescript-eslint/no-unsafe-return': 'error',
       '@typescript-eslint/no-unsafe-unary-minus': 'error',
-      '@typescript-eslint/no-useless-template-literals': 'error',
       'no-throw-literal': 'off',
       '@typescript-eslint/only-throw-error': 'error',
       'prefer-promise-reject-errors': 'off',