+++ packages/typescript-estree/tests/lib/semanticInfo-singleRun.test.ts
@@ -1,4 +1,4 @@
-import glob = require('glob');
+import * as glob from 'glob';
 import * as path from 'path';
 
 import { createProgramFromConfigFile as createProgramFromConfigFileOriginal } from '../../src/create-program/useProvidedPrograms';